+++ ledger/light/proc/get_request_test.go
@@ -149,7 +149,7 @@ func TestGetRequest_Proceed(t *testing.T) {
 		assert.Error(t, err)
 		insError, ok := errors.Cause(err).(*payload.CodedError)
 		assert.True(t, ok)
-		assert.Equal(t, uint32(payload.CodeNotFound), insError.GetCode())
+		assert.Equal(t, payload.CodeNotFound, insError.GetCode())
 	})
 
 	t.Run("Simple success", func(t *testing.T) {
